requirements.txt
openai-whisper==20230124
soundfile==0.12.1
SpeechRecognition==3.9.0
PyAudio==0.2.13
